Move and wrap up promos controller into admins module
