--- app/controllers/promos_controller.rb
@@ -1,2 +1,0 @@
-class PromosController < ApplicationController
-end
